Move `emscripten_atomic_wait_async` to atomic.h and add core testing
This API was orininally added for wasm workers but is not wasm worker
specific.

Followup to #20381.
