+++ system/include/emscripten/em_types.h
@@ -33,3 +33,16 @@ typedef void (*em_str_callback_func)(const char *);
 #define EM_TRUE 1
 #define EM_FALSE 0
 #define EM_UTF8 char
+
+#define EMSCRIPTEN_RESULT int
+
+#define EMSCRIPTEN_RESULT_SUCCESS              0
+#define EMSCRIPTEN_RESULT_DEFERRED             1
+#define EMSCRIPTEN_RESULT_NOT_SUPPORTED       -1
+#define EMSCRIPTEN_RESULT_FAILED_NOT_DEFERRED -2
+#define EMSCRIPTEN_RESULT_INVALID_TARGET      -3
+#define EMSCRIPTEN_RESULT_UNKNOWN_TARGET      -4
+#define EMSCRIPTEN_RESULT_INVALID_PARAM       -5
+#define EMSCRIPTEN_RESULT_FAILED              -6
+#define EMSCRIPTEN_RESULT_NO_DATA             -7
+#define EMSCRIPTEN_RESULT_TIMED_OUT           -8
