+++ system/include/emscripten/html5.h
@@ -70,19 +70,6 @@ extern "C" {
 #define EMSCRIPTEN_EVENT_CANVASRESIZED         37
 #define EMSCRIPTEN_EVENT_POINTERLOCKERROR      38
 
-#define EMSCRIPTEN_RESULT int
-
-#define EMSCRIPTEN_RESULT_SUCCESS              0
-#define EMSCRIPTEN_RESULT_DEFERRED             1
-#define EMSCRIPTEN_RESULT_NOT_SUPPORTED       -1
-#define EMSCRIPTEN_RESULT_FAILED_NOT_DEFERRED -2
-#define EMSCRIPTEN_RESULT_INVALID_TARGET      -3
-#define EMSCRIPTEN_RESULT_UNKNOWN_TARGET      -4
-#define EMSCRIPTEN_RESULT_INVALID_PARAM       -5
-#define EMSCRIPTEN_RESULT_FAILED              -6
-#define EMSCRIPTEN_RESULT_NO_DATA             -7
-#define EMSCRIPTEN_RESULT_TIMED_OUT           -8
-
 #define EMSCRIPTEN_EVENT_TARGET_INVALID        0
 #define EMSCRIPTEN_EVENT_TARGET_DOCUMENT       ((const char*)1)
 #define EMSCRIPTEN_EVENT_TARGET_WINDOW         ((const char*)2)
