+++ test/atomic/test_wait_async.out
@@ -1,0 +1,12 @@
+main: creating worker
+Async waiting on address with unexpected value should return 'not-equal'
+Waiting on address with unexpected value should return 'not-equal' also if timeout==0
+Waiting for 0 milliseconds should return 'timed-out'
+Waiting for >0 milliseconds should return 'ok' (but successfully time out in first call to asyncWaitFinished)
+Waiting for infinitely long should return 'ok' (and return 'ok' in second call to asyncWaitFinished)
+main: asyncWaitShouldTimeout
+worker_main
+worker: addr = 1234
+worker: notify async waiting main thread
+main: asyncWaitFinished
+test finished
