+++ test/test_core.py
@@ -2936,13 +2936,18 @@ The current type of b is: 9
   @node_pthreads
   def test_pthread_wait32_notify(self):
     self.set_setting('EXIT_RUNTIME')
-    self.do_run_in_out_file_test(test_file('wasm_worker/wait32_notify.c'))
+    self.do_run_in_out_file_test(test_file('atomic/test_wait32_notify.c'))
 
   @node_pthreads
   @no_wasm2js('https://github.com/WebAssembly/binaryen/issues/5991')
   def test_pthread_wait64_notify(self):
     self.set_setting('EXIT_RUNTIME')
-    self.do_run_in_out_file_test(test_file('wasm_worker/wait64_notify.c'))
+    self.do_run_in_out_file_test(test_file('atomic/test_wait64_notify.c'))
+
+  @node_pthreads
+  def test_pthread_wait_async(self):
+    self.set_setting('PROXY_TO_PTHREAD')
+    self.do_run_in_out_file_test(test_file('atomic/test_wait_async.c'))
 
   def test_tcgetattr(self):
     self.do_runf(test_file('termios/test_tcgetattr.c'), 'success')
@@ -9941,7 +9946,7 @@ NODEFS is no longer included by default; build with -lnodefs.js
   @node_pthreads
   def test_wasm_worker_wait_async(self):
     self.prep_wasm_worker_in_node()
-    self.do_runf(test_file('wasm_worker/wait_async.c'), emcc_args=['-sWASM_WORKERS'])
+    self.do_runf(test_file('atomic/test_wait_async.c'), emcc_args=['-sWASM_WORKERS'])
 
 
 # Generate tests for everything
