+++ test/wasm_worker/post_function.c
@@ -1,4 +1,5 @@
 #include <emscripten.h>
+#include <emscripten/console.h>
 #include <emscripten/wasm_worker.h>
 #include <assert.h>
